pyproject.toml
[tool.poetry]
name = "anki-patcher"
version = "0.1.0"
description = ""
authors = ["Alexander <alexanderweichart@fastmail.com>"]
readme = "README.md"
packages = [{ include = "anki_patcher" }]

[tool.poetry.scripts]
anki-patcher = "anki_patcher.main:main"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
argparse = "^1.4.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
